+++ packages/pwa-kit-mcp/src/server/server.js
@@ -15,7 +15,8 @@ import {
     TestWithPlaywrightTool,
     CreateNewPageTool,
     InstallAgentRulesTool,
-    ExploreCommerceAPITool
+    ExploreCommerceAPITool,
+    HooksRecommendationTool
 } from '../tools'
 import {Telemetry} from '../utils/telemetry'
 import {PWA_KIT_DESCRIPTIVE_NAME} from '../utils/constants'
@@ -70,6 +71,7 @@ class PwaStorefrontMCPServerHighLevel {
         this.createAppGuidelinesTool = new CreateAppGuidelinesTool()
         this.testWithPlaywrightTool = new TestWithPlaywrightTool()
         this.exploreCommerceAPITool = new ExploreCommerceAPITool()
+        this.hooksRecommendationTool = new HooksRecommendationTool()
         this.setupTools()
     }
 
@@ -119,6 +121,12 @@ class PwaStorefrontMCPServerHighLevel {
             this.exploreCommerceAPITool.description,
             this.exploreCommerceAPITool.inputSchema,
             this.exploreCommerceAPITool.handler
+        )
+        this.server.tool(
+            this.hooksRecommendationTool.name,
+            this.hooksRecommendationTool.description,
+            this.hooksRecommendationTool.inputSchema,
+            this.hooksRecommendationTool.handler
         )
     }